+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.2.0
+
+Released May 10, 2024.
+
+- Fixed the cone angle being wrong when using DrawWireCone.
+- Added a version of DrawWireEllipsoid to specify the ellipsoid using a rotation.
+- Added a version of DrawWireEllipticCone to specify using angles instead of a vector basis.
+
 ## v0.1.0
 
 Released May 9, 2024.